:construction: Added useStorageConfig custom hook
